adds error handling
refactors user credential handling and adds error handling to it
